requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.10
astroid==2.1.0
attrs==19.3.0
autopep8==1.4.3
Babel==2.8.0
black==19.10b0
certifi==2019.3.9
chardet==3.0.4
click==7.1.2
colorama==0.4.3
defusedxml==0.5.0
distlib==0.3.1
Django==2.2
django-allauth==0.39.1
django-bootstrap3==14.1.0
django-crispy-forms==1.7.2
django-otp==0.9.3
django-phone-login==1.1.3
django-phone-verify==2.0.0
django-phonenumber-field==1.1.0
django-rest-framework==0.1.0
django-sendsms==0.2.3
django-tinymce4-lite==1.7.4
djangorestframework==3.11.0
filelock==3.0.12
Flask==0.12.2
future==0.18.2
geographiclib==1.50
geopy==2.0.0
googlemaps==4.4.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==1.7.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsmin==2.2.2
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
nexmo==2.4.0
numpy==1.19.1
oauthlib==3.0.1
pathspec==0.8.0
pefile==2019.4.18
pep8==1.7.1
phonenumbers==8.12.7
phonenumberslite==8.12.7
Pillow==6.2.0
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pygame==1.9.6
pyinstall==0.1.4
PyInstaller==3.6
PyJWT==1.7.1
pylint==2.2.2
pyotp==2.4.0
python-dotenv==0.14.0
python3-openid==3.1.0
pytz==2018.9
pywin32-ctypes==0.2.0
qrcode==6.1
regex==2020.7.14
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.1
twilio==6.44.1
typed-ast==1.1.1
urllib3==1.24.2
virtualenv==20.0.27
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1
wrapt==1.11.0
zipp==3.1.0
